+++ src/ui-win/anim/AnimTransView.cpp
@@ -33,7 +33,7 @@
 #include "ui-win/Utils.h"
 #include "ui-win/AnimCalc.h"
 #include "resource/ResHelper.h"
-#include "ui-win/Factory.h"
+#include "presenters/Datarefs.h"
 
 namespace ui {
 namespace win {
@@ -57,8 +57,13 @@ namespace win {
                         break;
                     case CHK_LOOP: setLoopEnable();
                         break;
-                    case BTN_DATAREF: Factory::showNotImplemented();;
+                    case BTN_DATAREF: {
+                        MSTR str;
+                        Utils::getText(cEditDataRef, str);
+                        cEditDataRef->SetText(presenters::Datarefs::selectData(str));
+                        setDataref();
                         break;
+                    }
                     case BTN_REVERSE_VALUE: reverseValues();
                         break;
                     case BTN_CALC_VALUE: calculateValues();
@@ -71,7 +76,7 @@ namespace win {
             }
             case WM_CUSTEDIT_ENTER: {
                 switch (LOWORD(wParam)) {
-                    case EDIT_DATAREF: setDrft();
+                    case EDIT_DATAREF: setDataref();
                         break;
                     default: break;
                 }
@@ -333,7 +338,7 @@ namespace win {
         mData.saveToNode();
     }
 
-    void AnimTransView::setDrft() {
+    void AnimTransView::setDataref() {
         mData.mDataref = sts::toMbString(Utils::getText(cEditDataRef));
         setDataRefValueAsToolType();
         mData.saveToNode();