+++ src/ui-win/anim/AnimTransView.h
@@ -76,7 +76,7 @@ namespace win {
         void setEnable();
         void setReverseEnable();
         void setLoopEnable();
-        void setDrft();
+        void setDataref();
         void loopSpnChanged();
         void valueSpnChanged();
         void selectionChanged();