+++ src/ui-win/light/LightSpillCust.cpp
@@ -36,7 +36,7 @@
 #include "resource/resource.h"
 #include "ui-win/Utils.h"
 #include "resource/ResHelper.h"
-#include "ui-win/Factory.h"
+#include "presenters/Datarefs.h"
 
 namespace ui {
 namespace win {
@@ -69,7 +69,12 @@ namespace win {
             case WM_COMMAND: {
                 switch (LOWORD(wParam)) {
                     case IDC_BTN_DATAREF: {
-                        Factory::showNotImplemented();;
+                        MSTR str;
+                        Utils::getText(theDlg->cEdtDataRef, str);
+                        str = presenters::Datarefs::selectData(str);
+                        theDlg->cEdtDataRef->SetText(str);
+                        theDlg->mData->setDataRef(xobj::fromMStr(str));
+                        theDlg->eventParamChanged(true);
                         break;
                     }
                     default: break;