+++ src/ui-win/manip/ManipAttrDragAxisPix.cpp
@@ -38,7 +38,7 @@
 #include "resource/resource.h"
 #include "common/Logger.h"
 #include "resource/ResHelper.h"
-#include "ui-win/Factory.h"
+#include "presenters/Datarefs.h"
 
 namespace ui {
 namespace win {
@@ -71,7 +71,12 @@ namespace win {
             case WM_COMMAND: {
                 switch (LOWORD(wParam)) {
                     case BTN_DATAREF: {
-                        Factory::showNotImplemented();
+                        MSTR str;
+                        Utils::getText(theDlg->cEdtDataRef, str);
+                        str = presenters::Datarefs::selectData(str);
+                        theDlg->cEdtDataRef->SetText(str);
+                        theDlg->mData.setDataref(xobj::fromMStr(str));
+                        theDlg->save();
                         break;
                     }
                     case CMB_CURSOR: {